+++ src/client.rs
@@ -82,11 +82,19 @@ impl<'sto> Client<'sto> {
             .send(self, AlControl::reset())
             .await?;
 
-        // Clear FMMUs. FMMU memory section is 0xff (255) bytes long - see ETG1000.4 Table 57
-        self.blank_memory(RegisterAddress::Fmmu0, 0xff).await?;
+        // Clear FMMUs - see ETG1000.4 Table 57
+        // Some devices aren't able to blank the entire region.
+        for fmmu_idx in 0..16 {
+            self.blank_memory(RegisterAddress::fmmu(fmmu_idx), 0x10)
+                .await?;
+        }
 
-        // Clear SMs. SM memory section is 0x7f bytes long - see ETG1000.4 Table 59
-        self.blank_memory(RegisterAddress::Sm0, 0x7f).await?;
+        // Clear SMs - see ETG1000.4 Table 59
+        // Some devices aren't able to blank the entire region.
+        for sm_idx in 0..16 {
+            self.blank_memory(RegisterAddress::sync_manager(sm_idx), 0x8)
+                .await?;
+        }
 
         // Set DC control back to EtherCAT
         self.blank_memory(